change file names to batches and batch-builders
